requirements.txt
bcrypt==4.3.0
cachetools==5.5.2
cryptography==44.0.2
DateTime==5.5
keras==3.9.2
matplotlib==3.10.1
numpy==2.1.3
pandas==2.2.3
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
scikit-learn==1.6.1
scipy==1.15.2
streamlit==1.45.0
tensorflow==2.19.0
yfinance==0.2.59
python-dotenv==1.1.0